FIX: calculation projection

+++ src/calculate_map2d.c
@@ -10,7 +10,7 @@
 
 #include "my_struct.h"
 
-static sfVector2i project_iso_point(int x, int y, float z, map_t *map, sfVector2u *size)
+static sfVector2i project_iso_point(float x, float y, float z, map_t *map, sfVector2u *size)
 {
     sfVector2i point;
     float rx = cos(map->delta1 * M_PI / 180) * x - sin(map->delta1 * M_PI / 180) * y;
